base de datos lista conexion y desconexion
